backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
Django==4.2.29
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
geopandas==1.0.1
numpy==2.0.2
packaging==26.0
pandas==2.3.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
pyogrio==0.11.1
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
shapely==2.0.7
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
